Add Response.Content and Response.SetContent.
Now the Content object takes over the responsibility of downloading
the body content for the content scans.

+++ classify.go
@@ -137,7 +137,7 @@ func handleClassification(w http.ResponseWriter, r *http.Request) {
 	}
 
 	result.Categories = scores
-	logLine := logAccess(req, resp, len(content), modified, "", tally, scores, ACLActionRule{Action: "classify"}, "", nil, nil)
+	logLine := logAccess(req, resp, int64(len(content)), modified, "", tally, scores, ACLActionRule{Action: "classify"}, "", nil, nil)
 	if r.URL.Path == "/classify/verbose" {
 		result.LogLine = logLine
 	}
